// packages/sensinact-mapping-wizard/src/generated/FeaturePath.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, EStructuralFeature, BasicEObject, EClass } from '@emfts/core';

/**
 * FeaturePath
 * @generated
 */
export interface FeaturePath extends EObject {
  segments: EStructuralFeature[];
  collectionIndex?: number;
  label?: string;

}