// packages/sensinact-mapping-wizard/src/generated/SensorMappingSetup.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, EClass, BasicEObject, EStructuralFeature } from '@emfts/core';
import type { NameSource } from './NameSource';
import type { FeaturePath } from './FeaturePath';
import type { FriendlyNameSource } from './FriendlyNameSource';
import type { TimestampChoice } from './TimestampChoice';
import type { LocationChoice } from './LocationChoice';
import type { Measurement } from './Measurement';

/**
 * SensorMappingSetup
 * @generated
 */
export interface SensorMappingSetup extends EObject {
  mappingId: string;
  sensorClass: EClass;
  nameSource: NameSource;
  namePath?: FeaturePath;
  nameFallback?: string;
  friendlyNameSource: FriendlyNameSource;
  friendlyNamePath?: FeaturePath;
  friendlyName?: string;
  timestamp: TimestampChoice;
  location?: LocationChoice;
  measurements: Measurement[];

}