// packages/sensinact-mapping-wizard/src/generated/NameSource.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

/**
 * NameSource
 * @generated
 */
export const NameSource = {
  FROM_FIELD: 'FROM_FIELD',
  STATIC: 'STATIC'
} as const;

export type NameSource = typeof NameSource[keyof typeof NameSource];

/**
 * Get enum value by name
 */
export function getNameSource(name: string): NameSource | undefined {
  return (NameSource as Record<string, string>)[name] as NameSource | undefined;
}

/**
 * Get enum value by literal
 */
export function getNameSourceByLiteral(literal: string): NameSource | undefined {
  if (literal === 'FROM_FIELD') return NameSource.FROM_FIELD;
  if (literal === 'STATIC') return NameSource.STATIC;
  return undefined;
}